+++ Sources/SuperwallKit/Paywall/Presentation/PaywallCloseReason.swift
@@ -11,11 +11,11 @@ import Foundation
 /// interaction or because another paywall will show.
 @objc(SWKPaywallCloseReason)
 public enum PaywallCloseReason: Int, Codable, Equatable, Sendable {
-  /// The paywall was closed by system logic, either after a purchase, because
-  /// a deeplink was presented, close button pressed, etc.
+  /// The paywall was closed by system logic, either after a purchase or because
+  /// a deeplink was presented.
   case systemLogic
 
-  /// The paywall was automatically closed becacuse another paywall will show.
+  /// The paywall was automatically closed because another paywall will show.
   ///
   /// This prevents ``Superwall/register(event:params:handler:feature:)`` `feature`
   /// block from executing on dismiss of the paywall, because another paywall is set to show
@@ -27,6 +27,19 @@ public enum PaywallCloseReason: Int, Codable, Equatable, Sendable {
   /// handler will be called. If it's for a non-gated paywall, the feature block will be called.
   case webViewFailedToLoad
 
+  /// The paywall was closed because the user tapped the close button or dragged to dismiss.
+  case manualClose
+
   /// The paywall hasn't been closed yet.
   case none
+
+  /// Indicates whether the paywall state publisher should receive a completion.
+  var stateShouldComplete: Bool {
+    switch self {
+    case .forNextPaywall:
+      return false
+    default:
+      return true
+    }
+  }
 }