+++ Sources/SuperwallKit/Paywall/View Controller/PaywallViewController.swift
@@ -131,6 +131,13 @@ public class PaywallViewController: UIViewController, LoadingDelegate {
   private var presentationDidFinishPrepare = false
   private var didCallDelegate = false
 
+  /// `true` if there's a survey to complete and the paywall is displayed in a modal style.
+  private var didDisableSwipeForSurvey = false
+
+  /// The presenting view controller, saved for presenting surveys from when
+  /// the view disappears.
+  var internalPresentingViewController: UIViewController?
+
   private unowned let factory: TriggerSessionManagerFactory
   private unowned let storage: Storage
   private unowned let deviceHelper: DeviceHelper
@@ -230,7 +237,8 @@ public class PaywallViewController: UIViewController, LoadingDelegate {
   /// Triggered by user closing the paywall when the webview hasn't loaded.
   @objc private func forceClose() {
     dismiss(
-      result: .declined
+      result: .declined,
+      closeReason: .systemLogic
     ) { [weak self] in
       guard let self = self else {
         return
@@ -566,6 +574,16 @@ extension PaywallViewController: SWWebViewDelegate {
   }
 }
 
+// MARK: - UIAdaptivePresentationControllerDelegate
+extension PaywallViewController: UIAdaptivePresentationControllerDelegate {
+  public func presentationControllerDidAttemptToDismiss(_ presentationController: UIPresentationController) {
+    dismiss(
+      result: .declined,
+      closeReason: .manualClose
+    )
+  }
+}
+
 // MARK: - PaywallMessageHandlerDelegate
 extension PaywallViewController: PaywallMessageHandlerDelegate {
   func eventDidOccur(_ paywallEvent: PaywallWebEvent) {
@@ -598,7 +616,8 @@ extension PaywallViewController: PaywallMessageHandlerDelegate {
 
   func openDeepLink(_ url: URL) {
     dismiss(
-      result: .declined
+      result: .declined,
+      closeReason: .systemLogic
     ) { [weak self] in
       self?.eventDidOccur(.openedDeepLink(url: url))
       UIApplication.shared.open(url)
@@ -633,6 +652,16 @@ extension PaywallViewController {
     guard presentationWillPrepare else {
       return
     }
+    if let survey = paywall.survey,
+      survey.hasSeenSurvey(storage: storage) == false,
+      modalPresentationStyle == .formSheet
+      || modalPresentationStyle == .pageSheet
+      || modalPresentationStyle == .popover,
+      presentationController?.delegate == nil {
+      didDisableSwipeForSurvey = true
+      presentationController?.delegate = self
+      isModalInPresentation = true
+    }
     addShimmerView(onPresent: true)
 
     view.alpha = 1.0
@@ -650,9 +679,10 @@ extension PaywallViewController {
     presentationWillPrepare = false
   }
 
+
   public override func viewDidAppear(_ animated: Bool) {
     super.viewDidAppear(animated)
-
+    internalPresentingViewController = presentingViewController
     presentationDidFinish()
   }
 
@@ -713,22 +743,39 @@ extension PaywallViewController {
 
   func dismiss(
     result: PaywallResult,
-    closeReason: PaywallCloseReason = .systemLogic,
+    closeReason: PaywallCloseReason,
     completion: (() -> Void)? = nil
   ) {
     dismissCompletionBlock = completion
     paywallResult = result
     paywall.closeReason = closeReason
 
-    if let delegate = delegate {
-      didCallDelegate = true
-      delegate.didFinish(
-        paywall: self,
-        result: result,
-        shouldDismiss: true
-      )
-    } else {
-      dismiss(animated: presentationIsAnimated)
+    let isDeclined = paywallResult == .declined
+    let isManualClose = closeReason == .manualClose
+
+    func dismissView() {
+      if let delegate = delegate {
+        didCallDelegate = true
+        delegate.didFinish(
+          paywall: self,
+          result: result,
+          shouldDismiss: true
+        )
+      } else {
+        dismiss(animated: presentationIsAnimated)
+      }
+    }
+
+    SurveyManager.presentSurveyIfAvailable(
+      paywall.survey,
+      using: self,
+      loadingState: loadingState,
+      paywallIsManuallyDeclined: isDeclined && isManualClose,
+      isDebuggerLaunched: request?.flags.isDebuggerLaunched == true,
+      paywallInfo: info,
+      storage: storage
+    ) {
+      dismissView()
     }
   }
 
@@ -757,7 +804,7 @@ extension PaywallViewController {
       )
     }
 
-    if paywall.closeReason == .systemLogic {
+    if paywall.closeReason.stateShouldComplete {
       paywallStateSubject?.send(completion: .finished)
       paywallStateSubject = nil
     }
@@ -765,6 +812,13 @@ extension PaywallViewController {
     // Reset state
     Superwall.shared.destroyPresentingWindow()
     GameControllerManager.shared.clearDelegate(self)
+
+
+    if didDisableSwipeForSurvey {
+      presentationController?.delegate = nil
+      isModalInPresentation = false
+      didDisableSwipeForSurvey = false
+    }
 
     paywallResult = nil
     cache?.activePaywallVcKey = nil